Restructure directory/module layout
