+++ src/shared/mod.rs
@@ -1,2 +1,3 @@
+pub mod ffi;
 pub mod interface;
 pub mod state;
