--- src/wayland/dispatch/mod.rs
@@ -1,6 +1,0 @@
-pub mod buffer;
-pub mod empty;
-pub mod keyboard;
-pub mod lock;
-pub mod output;
-pub mod registry;
