+++ src/wayland/graphics/mod.rs
@@ -1,0 +1,3 @@
+pub mod buffer;
+pub mod output;
+pub mod renderer;
