+++ src/wayland/input/mod.rs
@@ -1,0 +1,1 @@
+pub mod keyboard;
