+++ src/wayland/registry/mod.rs
@@ -1,0 +1,1 @@
+pub mod registry;
